requirements.txt
Appium-Python-Client==0.49
et-xmlfile==1.0.1
jdcal==1.4.1
numpy==1.11.1
opencv-python==3.4.2.16
openpyxl==2.6.4
selenium==3.141.0
urllib3==1.25.9
matplotlib==2.2.5
opencv-contrib-python==3.4.2.16